+++ locale_posix.go
@@ -1,5 +1,5 @@
 // +build aix dragonfly freebsd hurd illumos linux nacl netbsd openbsd plan9 solaris zos
-// +build !unit_test
+// +build !tests
 
 package locale
